% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{sa_toggle}
\alias{sa_toggle}
\alias{update_sa_toggle}
\title{Toggle input}
\usage{
sa_toggle(
  inputId,
  label,
  value = FALSE,
  true = "Yes",
  false = "No",
  color = "blue",
  text_color = "black",
  width = NULL
)

update_sa_toggle(
  session,
  inputId,
  label = NULL,
  value = NULL,
  true = NULL,
  false = NULL
)
}
\arguments{
\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{logical, default toggle state.}

\item{true}{character, left label for \code{TRUE} state.}

\item{false}{character, right label for \code{FALSE} state.}

\item{color}{active state color.}

\item{text_color}{color of left and right value label text.}

\item{width}{The width of the input, e.g. '400px', or '100%'; see
\code{shiny::validateCssUnit}.}

\item{session}{The session object passed to function given to shinyServer.}
}
\value{
HTML element that can be added to a UI definition.
}
\description{
A Shiny binary toggle input.
}
\details{
This input returns \code{TRUE} or \code{FALSE} similar to
\code{shiny::checkboxInput}.
}
\examples{
if (interactive()) {
library(shiny)

ui <- function(request) {
fluidPage(
  use_access(),
  fluidRow(bookmarkButton()),
  fluidRow(
    column(3,
      actionButton("btn1", "Set to TRUE"),
      actionButton("btn2", "Change label"),
      sa_toggle(
        "toggle1", "Show text A", FALSE, "On", "Off",
        "firebrick", "white", 200
      ),
      sa_toggle(
        "toggle2", "Show text B", TRUE, "On", "Off",
        "yellowgreen", "black"
      )
    ),
    column(9, textOutput("txt1"), textOutput("txt2"))
  )
)
}

server <- function(input, output, session) {
  output$txt1 <- renderText(paste("Toggle input 1 is set to", input$toggle1))
  output$txt2 <- renderText(paste("Toggle input 2 is set to", input$toggle2))
  observeEvent(input$btn1, {
    update_sa_toggle(session, "toggle1", value = TRUE)
  })
  observeEvent(input$btn2, {
    update_sa_toggle(session, "toggle1", label = "Show text A 2",
                     true = "Yes", false = "No")
  })
}

shinyApp(ui, server, enableBookmarking = "url")

}
}
