% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_lang}
\alias{set_lang}
\title{Add HTML language attribute}
\usage{
set_lang(session, lang = "en")
}
\arguments{
\item{session}{Shiny session.}

\item{lang}{character, language abbreviation.}
}
\value{
nothing returned
}
\description{
Add an HTML language attribute to an app.
}
\details{
You can inspect the element in your browser to verify. The \code{<html>} tag
will include \code{lang="en"} by default.
}
\examples{
if(interactive()){
  library(shiny)

  shinyApp(
    ui = fluidPage(
      use_access(),
      h3("App title")
    ),
    server = function(input, output, session){
      set_lang(session, "en")
    }
  )
}
}
