% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_roles}
\alias{set_roles}
\title{Add role attributes}
\usage{
set_roles(session)
}
\arguments{
\item{session}{Shiny session.}
}
\value{
nothing returned
}
\description{
Add role language attribute to key Shiny widgets.
}
\details{
You can inspect the element in your browser to verify the addition of role
attributes for the following Shiny elements.

\itemize{
  \item \code{tabsetPanel}
  \item \code{tabPanel}
}
}
\examples{
if(interactive()){
  library(shiny)

  shinyApp(
    ui = fluidPage(
      use_access(),
      h3("App title"),
      tabsetPanel(
        tabPanel("panel1", "Content"),
        tabPanel("panel2", "Content")
      ),
      tabsetPanel(
        tabPanel("panel1", "Content"),
        tabPanel("panel2", "Content"),
        type = "pills", id = "tsp2"
      )
    ),
    server = function(input, output, session){
      set_roles(session)
    }
  )
}
}
