% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyaccess.R
\docType{package}
\name{shinyaccess}
\alias{shinyaccess}
\title{shinyaccess: Accessibility in Shiny}
\description{
This package provides functions and other resources for enhancing web
accessibility in Shiny apps.
}
