% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyaccess.R
\name{use_access}
\alias{use_access}
\title{Initialize shinyaccess}
\usage{
use_access()
}
\value{
nothing returned
}
\description{
Initialize shinyaccess in Shiny UI.
}
\examples{
if(interactive()){
  library(shiny)

  shinyApp(
    ui = fluidPage(
      use_access(),
      h3("App title")
    ),
    server = function(input, output, session){
      set_lang(session, "en")
    }
  )
}
}
