% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm_checkboxgroup.R
\name{dsmCheckboxGroupInput}
\alias{dsmCheckboxGroupInput}
\title{DSM checkbox group input}
\usage{
dsmCheckboxGroupInput(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  values = choices,
  width = NULL
)
}
\arguments{
\item{inputId}{input ID.}

\item{label}{text label.}

\item{choices}{selection choices.}

\item{selected}{not in use}

\item{inline}{not in use}

\item{width}{not in use}

\item{value}{defaults to \code{choices}.}
}
\value{
A list of HTML elements that can be added to a UI definition.
}
\description{
DSM checkbox group input
}
\examples{
if (interactive()) {

ui <- fluidPage(
  fluidRow(
    column(3,
      checkboxGroupInput("cb1", "Shiny1 - Variables to show:",
                         list("Cylinders" = "cyl", "Transmission" = "am", "Gears" = "gear")),
      tableOutput("data1")
    ),
    column(3,
      checkboxGroupInput("cb2", "Shiny2 - Variables to show:",
                         list("Cylinders" = "cyl", "Transmission" = "am", "Gears" = "gear")),
      tableOutput("data2")
    ),
    column(3,
      dsmCheckboxGroupInput("cb3", "DSM1 - Variables to show:",
                            c("Cylinders", "Transmission", "Gears"),
                            selected = c("cyl", "am"),
                            values = c("cyl", "am", "gear")),
      tableOutput("data3")
    ),
    column(3,
      dsmCheckboxGroupInput("cb4", "DSM2 - Variables to show:",
                            c("Cylinders", "Transmission", "Gears"),
                            selected = c("cyl", "am"),
                            values = c("cyl", "am", "gear")),
      tableOutput("data4")
    )
  )
)

server <- function(input, output, session) {
  output$data1 <- renderTable({
    mtcars[, c("mpg", input$cb1), drop = FALSE]
  }, rownames = TRUE)

  output$data2 <- renderTable({
    mtcars[, c("mpg", input$cb2), drop = FALSE]
  }, rownames = TRUE)
  output$data3 <- renderTable({
    mtcars[, c("mpg", input$cb3), drop = FALSE]
  }, rownames = TRUE)

  output$data4 <- renderTable({
    mtcars[, c("mpg", input$cb4), drop = FALSE]
  }, rownames = TRUE)
}

shinyApp(ui, server)

}
}
